Adjusts in migrations

+++ src/main/resources/db/migration/V3__create_table_filme_genero.sql
@@ -1,5 +1,5 @@
 CREATE TABLE filme_genero (
-    filme_id BIGINT NOT NULL REFERENCES filmes(id),
-    genero_id BIGINT NOT NULL REFERENCES generos(id),
+    filme_id INT NOT NULL REFERENCES filmes(id),
+    genero_id INT NOT NULL REFERENCES generos(id),
     PRIMARY KEY (filme_id, genero_id)
 );
